backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0
python-pptx==0.6.23
Pillow>=10.2.0
google-generativeai>=0.8.3
python-dotenv==1.0.0
Werkzeug==3.0.1

